Move errors to util
